Add CREDITS.md, comment fixes.
The comments inside of source files are now less pretentious (tm),
saying "source" with a link to the repo rather than "author" with my
name as well as a link. For credits CREDITS.md will be used instead.
While I was at it I also cleaned up the Makefile a tiny bit.